Move 'StateMachineTest' under 'test-utils' so that it can be visible by both tests and integration tests; Made 'SingleThreadedStateMachineManager' non internal so its hooks are visible outside 'node' module.
